requirements.txt
asgiref==3.8.1
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==44.0.0
distlib==0.3.9
Django==5.1.4
django-allauth==65.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
exceptiongroup==1.2.2
Faker==33.1.0
filelock==3.16.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyproject-api==1.8.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tomli==2.2.1
tox==4.23.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
whitenoise==6.8.2
